DOC: CCT: missing encoded result for 8
In the description for "HammingCodes: Integer to Encoded Binary", the problem description gives the encoding pattern for 8, but not its encoded form.  However, for 21 the description gives the encoding pattern as well as the encoded form.  Present the encoded form for 8.  A tiny bit of rewrite to make both examples as separate sentences.