Update Documentation to Explain Args Usage in NS2
First, this repo does not seem to be where the site I'm reading is generated from; thus, this is not the documentation I saw.

Second, unless the tail function does something odd I have yet to see, when I attempt to run the NS2 code shown in this file it does not work.

The documentation does not explain how NS2 scopes the game arguments, I wasted 30 minutes trying to figure out why I couldn't get args to work as this file stated. It's explained in other files, but this one looks like it missed the updating.

Simple fix, I suspect this is just a minor oversight, code now includes an example of how to use it with ns2.