Significantly improve the speed of script creation
In my browser (Chrome 107.0.5304.107 Windows 10) this improves noop
script execution from 8281.063/sec to 13942.606/sec - a 68% speedup!

The 'ns' API has to be wrapped, so that functions can access the context
(primarily the WorkerScript) without leaking it to the player. Before,
this happened by walking the whole API tree and creating new wrapping
functions for every script, which is understandably expensive.

Now we take advantage of class #private fields to securely hold the
context, allowing the wrapping functions to exist on the prototype and
only need to be created once. These are widely available (>93% according
to caniuse) and work with Electron.