DOC: sleeve: get the Sleeve API to be linked in the list of APIs
The interface that defines the Sleeve object clashes with the interface that defines the Sleeve API.  The former is named Sleeve (uppercase), whereas the latter is named sleeve (lowercase).  API Extractor does not like this and inserts a link to the Sleeve object in place of where the Sleeve API should be.  A simple fix is to rename the interface for the Sleeve object as `SleeveObject`.