DOC: `ns.codingcontract.attempt()`: how to use optional parameter `opts`
Explain how to use the optional parameter `opts` of the function `ns.codingcontract.attempt()`.  It is not obvious that an object should be passed in, using the key/value format of the interface `CodingAttemptOptions`.  I have seen players passed in `true` for this fourth parameter.  The correct way to use the fourth parameter `opts` is to pass in an object along the key/value format of `CodingAttemptOptions`.  Also update the documentation of `opts` and the return value.