Revert "Revert "Add keycode table and switch to event.code in select places" (#224)"
The change did in fact reintroduce https://github.com/danielyxie/bitburner/issues/3592
The issue was very Mac-specific, which is why my testing didn't/couldn't
catch it.

Plan is to bring back the use of event.key eventually (to fix the issue
with different keyboard layouts), but only with a way to configure
hotkeys so that users can work around their platform-specific issues.

This reverts commit 70fadde222d7f79b3aa24fd3167d78669a63b169.