NETSCRIPT: Greatly speed scp() and write() of js by deferring RAM calculation
In the course of investigating a typo, I realized that there's no need
for ns.scp() and ns.write() to calculate ram at all. They can defer the
work until when the script is executed. The only visible artifact is
that RAM costs will be incorrect until the script is run; however, they
were already wrong in certain cases for scp anyway.