MULTS: Make BitNodeMultipliers.ServerMaxMoney more sensible
Before, a server's max money was effectively multiplied by both
ServerMaxMoney *and* ServerStartingMoney. This changes it so they both
act independently, while updating the values so that all servers will
continue to have the same amounts as before.

Players will see the new values in both the stats screen and through
code (getBitnodeMultipliers()), but the actual effect remains the same.

This leads to some surprising results (that are now more apparent), such
as ServerMaxMoney decreasing with a 2x exponent compared to starting
money in BN12. The values can be rebalanced later if desired.