New ns implementation using Proxy
This should be at least as fast as the previous method, since very few
objects are being created at script startup. Instead, we pay the cost of
wrapping incrementally, as functions are called. (But the wrapped
version is memoized.)

NS1 will be slow, since it traverses the entire tree and re-wraps
everything.