NETSCRIPT: More ns Proxy changes
This changes the handler implementation to be a class, instead of a
plain-old-object. This saves on the re-creation of all the functions
that make it up, and generally reduces the number of objects and
functions that need to be allocated for each layer.

The other change is to simplify the typing by removing ExternalAPI and
just using the API directly. The API was designed to be the External
API, so there's no need for the extra wrapper. Introducing the
GenericAPI type makes some things cleaner, but it required changing some
interfaces into types for TypeScript to be happy. (This apparently
wouldn't be needed on a newer version of TypeScript.)

Using the same benchmarker as #223, I'm seeing script creation of
55000/sec (!). The 2.2.0 approach was ~13000/sec. With Proxy, before
this PR, it was ~53000/sec. (Maybe slightly slower, but hard to be
sure.)