CORP: Revert mandatory coffeeparty
Steady-state Morale/Happiness/Energy drain is not a good gameplay
mechanic. It forces the player to do a tedious task for no real reason.
Worse, it disproportionally punishes new players, who already have an
uphill battle dealing with corp and trying to play it manually.

This goes back to the 2.1 model of steady-state morale increase. The
multiplier is much lower than it was then; for best results, coffeeparty
should still be used, but it no longer is required on a recurring basis.

This also fixes a bug where only revenue was being checked; now the
multiplier will properly decrease when profits are negative.