NETSCRIPT: Add growthAnalyzeCorrected + formulas version
ns.growthAnalyze is a trap for the unwary, since it returns inexact
results and presents a misleading view of how ns.grow works. However,
its usage is too baked in to various scripts to change, and the simpler
model has its place.

This adds ns.growthAnalyzeCorrected alongside, which is a wrapper to
numCyclesForGrowthCorrected. There's also a formulas version added, but
I believe it's important to have one in the base API so that players
without formulas access can still get exact results (with the trade-off
that they must prep their servers, first).