NETSCRIPT: Add "temporary" as a RunOption to run/exec/spawn
This creates the RunOptions interface, to pass as a parameter to
run()-type functions. For backwards compatibility, you can pass a thread
number or an options struct, and the struct has a "threads" member.
(This mirrors how JS itself has been extended.)

The new option is "temporary": It marks a RunningScript to be excluded
from the save. This is both a (significant) performance improvement, as
well as a useful semantic: There are lots of times where you have a
central "manager" script that you want to be restarted on load, but it
spawns helper or "worker" scripts that shouldn't be restarted.

It's possible to write your code so that the helpers will immediately
exit if they detect that they are running from a fresh load, but it's
both simpler, smaller, and more performant to never save them in the
first place.