NETSCRIPT: Add ramOverride as a RunOption
This lets you launch scripts with an arbitrary amount of RAM. If you use
this, it's up to you to ensure that the script has enough RAM to pass
the dynamic RAM check.

This is useful for large manager scripts where you might want to launch
them with a lower cost when you aren't using their expensive functions
(like corp). Also useful for a single HGW worker script that does a
similar thing for each operation, but wants to use 1.7GB for hack.

Testing revealed issues with `WorkerScript.ramUsage` (it was possible to
bypass the dynamic RAM check), so I removed it. There should be a
single source of truth.