Break the import cycle of Player
This reorders the dependency between Player (the global object) and
PersonObjects/Player/PlayerObject (the implementation) so that instead
of going Player -> PlayerObject, it now goes PlayerObject -> Player.

Since PlayerObject is a massive class, it depends on most things
througout the game, and at the same time most things depend on Player.
This massive dependency cycle permeated most things, and prevented a lot
of classes from being imported standalone without ordering issues.

Added 3 tests on important mid-level classes that used to fail to
import.