NETSCRIPT: Compiled modules will be even more shared
This creates greater memory savings by sharing compiled modules across
servers, when it is possible to do so. This introduces a small
user-visible change: Module-scoped variables will now be global to *all*
instances of that script, instead of per-server global.

This also revokes Blob URLs *immediately* after they are created. Blob
leakage should be a thing of the past!