UI: LogBox overhaul
This revamps a lot of the internals of LogBoxManager, although to users
it should act entirely the same, and the most visible difference will be
the use of icons instead of text for the buttons.

Added the ability to get/set the title, including having it be saved and
setting it to arbitrary ReactElements for the brave.

Added the ability to read the log window's size/position, as well as
determine if it is currently shown.