NETSCRIPT: Rebalance the cost of sing.getOwnedSourceFiles
This function (rightly, imo) does not actually require SF4. However,
since it currently has the SF4 cost multipliers applied to it, it is
largely moot since using it without SF4 requires 80GB.

This changes the RAM cost to be a constant 5GB. Ideally this wouldn't
live in the sing namespace at all (nothing that just returns information
should), but for compatibility that's how it rolls.