Reimplement #954
The changes made in #1083 were not compatible with #954. Making changes based on #954 was not practical due to the scope of changes so I reimplemented the functionality here.

The sorting method here was written to minimize iteration of the base array, which is why I took a different approach to the comparison logic.

* Added an isHacknetServer optional attribute to BaseServers
* HacknetServer has a truthy value added for the new attribute to make comparisons easier
* ActiveScriptsPage gets the server objects because they are needed for sorting. ActiveScriptsPage new passes the actual server object to the ServerAccordion, instead of just the hostname.
* Changed a temporary data structure from a regular object to a Map, so it can be keyed by actual server objects instead of hostnames.

Co-Authored-By: TheAimMan <147098375+TheAimMan@users.noreply.github.com>