Update safe loading method
Make the loading method work a bit better as far as ensuring bladeburner.contracts and bladeburner.operations are equal to the static versions and have no discrepancies regardless of possible savedata format differences. Also the loaded properties are all typechecked.
Also removed the `name` property from savedata on these objects, as it was no longer needed.