Remove non-working savegame conversion
This savegame conversion code doesn't work anyway, due to the use of getKeyList the load process will not load in the value for the old "blackops" property.
The only impact here is that savegames loaded into the new version while in the middle of bladeburner will have to restart the blackops from the beginning, which should only be a minor setback.