Two minorly breaking changes around script launch + refactoring.
This changes it so that "this" will be unbound when `main()` is called,
and also passes `ns.args` as additional arguments to the `main`
function.

Both of these are technically breaking changes, but in very minor ways:
Someone could be calling exported functions/variables with
`this.func()`, but given the simpler `func()` is available and works
universally for exported *and* unexported functions, it seems unlikely
that anyone would do this. It's also a very easy fix.

Similarly, adding new parameters to `main()` could break someone if they
were calling `main()` recursively and depending on them to be empty, as
they are now. Note that they will *still* be empty if the program is
called with no arguments, so this is also a tough needle to hit, in
exchange for better usability.

There is also some refactoring, to use `ScriptArg` consistently
everywhere.