Rename function
Renamed getAIMove to makeAIMove, to make it clear that it is not just for getting the move (like the getMove function) but is for actually applying the move to the board as well.
Not really necessary in this PR, just an editorial change since I was messing with that function anyway.