Fix all currently-known auto-infil bugs.
The largest part of this is a big refactor to how state is managed. All
useEffect()s have been yanked, in favor of directly assigning to global
state. This is simpler, probably faster, and guaranteed to work
synchronously. Various cases are guarded to ensure that we don't have
stale results lying around; it's not quite as elegant as before, but it
doesn't rely on React implementation details either.