UI: Add Fractional Digits option slider
This adds a slider to the Numeric Display options page that is tied to the fractionalDigits field in Settings.

Setting this to 0 works as intended, the formatter removes the seperator entirely and displays an integer instead of any amount of fractions behind it. Although, I did notice a minor edge case where the number sometimes displays a comma separated value like for example 1,000K when its really close to being a million, like 999995. I dont think this is caused by the changes applied in this fork..